% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whdxdata-package.R
\docType{package}
\name{whdxdata-package}
\alias{whdxdata-package}
\title{whdxdata: Water point exchange data.}
\description{
Data from the Water Point Data Exchange (WPDX) database.
}
\author{
\strong{Maintainer}: Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

}
\keyword{internal}
